updated timer functions variable type from int32 to uint32, because timers never gets negative value as base... also small code cleanup

+++ scripts/InstanceScripts/Instance_TheShatteredHalls.cpp
@@ -2304,10 +2304,10 @@ class WarbringerOmroggAI : public MoonScriptCreatureAI
 
 		MoonScriptCreatureAI*	mLeftHead;
 		MoonScriptCreatureAI*	mRightHead;
-		int32					mAggroShiftTimer;
-		int32					mBlastWaveTimer;
-		int32					mSpeechTimer;
-		int32					mSpeechId;
+		uint32					mAggroShiftTimer;
+		uint32					mBlastWaveTimer;
+		uint32					mSpeechTimer;
+		uint8					mSpeechId;
 		SpellDesc*				mBlastWave;
 };
